🐛 (#226) 코스 삭제 오류 해결과 삭제 위치 이동
