+++ app/member/course/page.tsx
@@ -12,13 +12,10 @@ import useToggle from "@/hooks/useToggle";
 import { useCourseStore } from "@/store/courseStore";
 import { InputFormData } from "@/components/modal/Modal.type";
 import { ShowCourseListDto } from "@/application/usecases/course/dto/ShowCourseListDto";
-import ScrollCardContainer from "@/components/CardContainer/ScrollCardContainer";
-import Icon from "@/components/Icon/Icon";
 
 export default function CoursePage() {
   const router = useRouter();
   const [isModalOpen, toggleModal] = useToggle(false);
-  const [isDeleteModalOpen, toggleDeleteModal] = useToggle(false);
   const { setName, setDate } = useCourseStore();
   const [courseList, setCourseList] = useState<ShowCourseListDto[]>([]);
   const [pastCourseList, setPastCourseList] = useState<ShowCourseListDto[]>([]);
@@ -29,19 +26,24 @@ export default function CoursePage() {
         const response = await fetch("/api/course", {
           credentials: "include",
         });
+
         if (!response.ok) {
           if (process.env.NODE_ENV === "development") {
             console.error("🚨 코스 목록 불러오기 실패");
           }
           return;
         }
+
+        const today = new Date();
+        today.setHours(0, 0, 0, 0);
+
         const data: ShowCourseListDto[] = await response.json();
 
         const currentCourse = data.filter(
-          (course) => new Date(course.date) > new Date()
+          (course) => new Date(course.date) >= today
         );
         const pastCourse = data.filter(
-          (course) => new Date(course.date) < new Date()
+          (course) => new Date(course.date) < today
         );
         setCourseList(currentCourse);
         setPastCourseList(pastCourse);
@@ -53,31 +55,6 @@ export default function CoursePage() {
     };
     fetchCourseList();
   }, []);
-
-  const handleCourseDelete = async (courseId: number) => {
-    try {
-      const response = await fetch("/api/course", {
-        method: "DELETE",
-        headers: {
-          "Content-Type": "application/json",
-        },
-        body: JSON.stringify({ courseId }),
-        credentials: "include",
-      });
-      if (response.ok) {
-        setCourseList((prev) =>
-          prev.filter((course) => course.id !== courseId)
-        );
-        setPastCourseList((prev) =>
-          prev.filter((course) => course.id !== courseId)
-        );
-      }
-    } catch (error) {
-      if (process.env.NODE_ENV === "development") {
-        console.error("🚨 코스 삭제 실패:", error);
-      }
-    }
-  };
 
   const handleConfirm = (inputFormData?: InputFormData) => {
     if (!inputFormData) return;
@@ -122,62 +99,29 @@ export default function CoursePage() {
       {courseList.length > 0 ? (
         <>
           <p>다가오는 코스</p>
-          <ScrollCardContainer variant="list">
-            {courseList.map((course) => (
-              <div key={course.id} className={styles.container}>
-                <div
-                  onClick={() => handleCourseClick(course)}
-                  className={styles.courseListView}
-                >
-                  <CourseListView {...course} isPast={false} />
-                </div>
-                <span className={styles.trash} onClick={toggleDeleteModal}>
-                  <Icon id="trash" />
-                </span>
-                <Modal
-                  contents={[
-                    {
-                      type: "textOnly",
-                      title: "정말로 코스를 삭제하시겠습니까?",
-                    },
-                  ]}
-                  onConfirm={() => handleCourseDelete(course.id)}
-                  onCancel={toggleDeleteModal}
-                  isOpen={isDeleteModalOpen}
-                  confirmText="삭제"
-                  cancelText="취소"
-                />
+          {courseList.map((course) => (
+            <div key={course.id} className={styles.container}>
+              <div
+                onClick={() => handleCourseClick(course)}
+                className={styles.courseListView}
+              >
+                <CourseListView {...course} isPast={false} />
               </div>
-            ))}
-          </ScrollCardContainer>
+            </div>
+          ))}
           {pastCourseList.length > 0 && (
             <>
               <p>종료된 코스</p>
-              <ScrollCardContainer variant="list">
-                {pastCourseList.map((course) => (
-                  <div key={course.id}>
-                    <div onClick={() => handleCourseClick(course)}>
-                      <CourseListView {...course} isPast={true} />
-                    </div>
-                    <span className={styles.trash} onClick={toggleDeleteModal}>
-                      <Icon id="trash" />
-                    </span>
-                    <Modal
-                      contents={[
-                        {
-                          type: "text",
-                          title: "정말로 코스를 삭제하시겠습니까?",
-                        },
-                      ]}
-                      onConfirm={() => handleCourseDelete(course.id)}
-                      onCancel={handleCancel}
-                      isOpen={isDeleteModalOpen}
-                      confirmText="삭제"
-                      cancelText="취소"
-                    />
+              {pastCourseList.map((course) => (
+                <div key={course.id} className={styles.container}>
+                  <div
+                    onClick={() => handleCourseClick(course)}
+                    className={styles.courseListView}
+                  >
+                    <CourseListView {...course} isPast={true} />
                   </div>
-                ))}
-              </ScrollCardContainer>
+                </div>
+              ))}
             </>
           )}
         </>
